+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.11.0
 * fix error in the javaFuncs helper
-* extend internal model with minItems and maxItems restrictions for arrays
+* extend internal model with minItems and maxItems restrictions for array properties
+* extend internal model with uniqueItems for array properties
 
 # 0.10.0
 * improvements for the Java bean template